Fix linux <-> windows path relocation issue
sometimes paths which did not contain the given root dir still made it past line 72; when relocation was attempted, the path wasn't fundamentally changed, BUT in the case of posix/windows relocation, the forwardslashes were changed to backslash. For repeated relocation of multiple folders to different places, this meant all paths not matched by the first relocation were all set to have backslash separators, meaning they are permanently unrelocatable by the subsequent relocations, since they are still looking for backslashes. The conditional that was supposed to detect this was broken, I fixed it